fix: removed redundant code and small refactors
