+++ src/Uno.UI.Toolkit/Storage/StorageFileHelper.wasm.cs
@@ -9,7 +9,7 @@ partial class StorageFileHelper
 {
 	private static async Task<bool> FileExistsInPackage(string fileName)
 	{
-		var assets = await AssetsManager.GetAssets(CancellationToken.None);
+		var assets = await AssetsManager._assets.Value;
 		return assets?.Contains(fileName) ?? false;
 	}
 }
